+++ astropy/io/fits/tests/test_convenience.py
@@ -23,3 +23,19 @@ class TestConvenience(FitsTestCase):
         with catch_warnings() as w:
             header = fits.getheader(self.data('test0.fits'))
             assert len(w) == 0
+
+    def test_fileobj_not_closed(self):
+        """
+        Tests that file-like objects are not closed after being passed
+        to convenenience functions.
+
+        Regression test for https://github.com/astropy/astropy/issues/5063
+        """
+
+        f = open(self.data('test0.fits'), 'rb')
+        data = fits.getdata(f)
+        assert not f.closed
+
+        f.seek(0)
+        header = fits.getheader(f)
+        assert not f.closed